refactor: sql queries

+++ kompeter-database/src/main/java/com/github/ragudos/kompeter/database/dao/inventory/PurchaseDao.java
@@ -10,6 +10,7 @@ package com.github.ragudos.kompeter.database.dao.inventory;
 import com.github.ragudos.kompeter.database.dto.enums.DiscountType;
 import com.github.ragudos.kompeter.database.dto.inventory.PurchaseDto;
 import java.io.IOException;
+import java.math.BigDecimal;
 import java.sql.SQLException;
 import java.sql.Timestamp;
 import java.util.List;
@@ -21,8 +22,8 @@ public interface PurchaseDao {
             Timestamp purchase_date,
             String purch_code,
             Timestamp deliverydate,
-            float vat_percentage,
-            float disc_val,
+            BigDecimal vat_percentage,
+            BigDecimal disc_val,
             DiscountType discountType)
             throws SQLException, IOException;
